+++ src/main/java/com/mcgamer199/luckyblock/customentity/EntityKiller.java
@@ -1,11 +1,11 @@
 package com.mcgamer199.luckyblock.customentity;
 
 import com.mcgamer199.luckyblock.engine.LuckyBlock;
+import com.mcgamer199.luckyblock.entity.CustomEntity;
 import org.bukkit.Material;
 import org.bukkit.entity.Item;
 import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.inventory.ItemStack;
-import com.mcgamer199.luckyblock.entity.CustomEntity;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -30,22 +30,22 @@ public class EntityKiller extends CustomEntity {
                 EntityKiller.items.remove(i);
                 i.remove();
             }
-        }, (long)time);
+        }, time);
     }
 
     protected void onDamage(EntityDamageEvent event) {
         int x;
         Item d;
         if (event.getDamage() < 100.0D) {
-            for(x = (int)((double)LuckyBlock.randoms.nextInt(5) + event.getDamage()); x > 0; --x) {
-                d = this.entity.getWorld().dropItem(this.entity.getLocation(), new ItemStack(Material.WOOL, 1, (short)14));
+            for (x = (int) ((double) LuckyBlock.randoms.nextInt(5) + event.getDamage()); x > 0; --x) {
+                d = this.entity.getWorld().dropItem(this.entity.getLocation(), new ItemStack(Material.WOOL, 1, (short) 14));
                 items.add(d);
                 d.setPickupDelay(2000);
                 this.remove(d, (LuckyBlock.randoms.nextInt(20) + 6) * 3);
             }
         } else {
-            for(x = LuckyBlock.randoms.nextInt(5) + 100; x > 0; --x) {
-                d = this.entity.getWorld().dropItem(this.entity.getLocation(), new ItemStack(Material.WOOL, 1, (short)14));
+            for (x = LuckyBlock.randoms.nextInt(5) + 100; x > 0; --x) {
+                d = this.entity.getWorld().dropItem(this.entity.getLocation(), new ItemStack(Material.WOOL, 1, (short) 14));
                 items.add(d);
                 d.setPickupDelay(2000);
                 this.remove(d, (LuckyBlock.randoms.nextInt(20) + 6) * 2);
