+++ src/main/java/com/mcgamer199/luckyblock/structures/BossDungeon.java
@@ -1,8 +1,8 @@
 package com.mcgamer199.luckyblock.structures;
 
+import com.mcgamer199.luckyblock.customentity.EntityGuardian;
 import com.mcgamer199.luckyblock.lb.LB;
 import com.mcgamer199.luckyblock.lb.LBType;
-import com.mcgamer199.luckyblock.customentity.EntityGuardian;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
@@ -25,26 +25,26 @@ public class BossDungeon extends LuckyDungeon {
         int r;
         int x;
         int z;
-        for(r = -5; r < 30; ++r) {
-            for(x = -5; x < 6; ++x) {
-                for(z = -5; z < 6; ++z) {
-                    block.getLocation().add((double)x, (double)r, (double)z).getBlock().setType(Material.BRICK);
+        for (r = -5; r < 30; ++r) {
+            for (x = -5; x < 6; ++x) {
+                for (z = -5; z < 6; ++z) {
+                    block.getLocation().add(x, r, z).getBlock().setType(Material.BRICK);
                 }
             }
         }
 
-        for(r = -3; r < 4; ++r) {
-            for(x = 1; x < 29; ++x) {
-                for(z = -3; z < 4; ++z) {
-                    block.getLocation().add((double)r, (double)x, (double)z).getBlock().setType(Material.AIR);
+        for (r = -3; r < 4; ++r) {
+            for (x = 1; x < 29; ++x) {
+                for (z = -3; z < 4; ++z) {
+                    block.getLocation().add(r, x, z).getBlock().setType(Material.AIR);
                 }
             }
         }
 
-        for(r = 0; r < 6; ++r) {
-            for(x = a * -1; x < a + 1; ++x) {
-                for(z = a * -1; z < a + 1; ++z) {
-                    block.getLocation().add((double)x, (double)(r + 30), (double)z).getBlock().setType(Material.BRICK);
+        for (r = 0; r < 6; ++r) {
+            for (x = a * -1; x < a + 1; ++x) {
+                for (z = a * -1; z < a + 1; ++z) {
+                    block.getLocation().add(x, r + 30, z).getBlock().setType(Material.BRICK);
                 }
             }
