+++ src/main/java/com/mcgamer199/luckyblock/structures/WoolHouse.java
@@ -20,9 +20,9 @@ public class WoolHouse extends Structure {
         int s = 0;
 
         int g;
-        for(g = 0; g < r; ++g) {
-            block.getLocation().add(0.0D, (double)g, 0.0D).getBlock().setType(Material.WOOL);
-            block.getLocation().add(0.0D, (double)g, 0.0D).getBlock().setData((byte)7);
+        for (g = 0; g < r; ++g) {
+            block.getLocation().add(0.0D, g, 0.0D).getBlock().setType(Material.WOOL);
+            block.getLocation().add(0.0D, g, 0.0D).getBlock().setData((byte) 7);
             ++s;
         }
 
@@ -31,14 +31,14 @@ public class WoolHouse extends Structure {
             ++g;
         }
 
-        byte b = (byte)this.random.nextInt(16);
+        byte b = (byte) this.random.nextInt(16);
         int k = this.random.nextInt(2) + 1;
 
-        for(int h = g; h > 0; --h) {
-            for(int x = g * -1; x < g - 1; ++x) {
-                for(int z = g * -1; z < g - 1; ++z) {
-                    block.getLocation().add((double)(x + 1), (double)s, (double)(z + 1)).getBlock().setType(Material.WOOL);
-                    block.getLocation().add((double)(x + 1), (double)s, (double)(z + 1)).getBlock().setData(b);
+        for (int h = g; h > 0; --h) {
+            for (int x = g * -1; x < g - 1; ++x) {
+                for (int z = g * -1; z < g - 1; ++z) {
+                    block.getLocation().add(x + 1, s, z + 1).getBlock().setType(Material.WOOL);
+                    block.getLocation().add(x + 1, s, z + 1).getBlock().setData(b);
                 }
             }
